Refactor: Move memory module to context adapters
- Moved memory/ module to context/adapters/memory/
- Updated redis_store.py location
- Fixed import paths and test references
- All tests passing with 307 passed, 0 failures
- Applied linting fixes with ruff
