+++ src/swe_ai_fleet/context/adapters/memory/cataloger.py
